src/make_us_rich/serving/requirements.txt
fastapi==0.73.0
minio==7.1.2
onnx==1.10.2
onnxruntime==v1.10.0
pandas==1.4.0
python-binance==v1.0.15
python-dotenv==0.19.2
scikit-learn==1.0.2
torch==1.10.1
uvicorn[standard]==0.15.0
